minor refactoring to make find token page support MC
